+++ variants/importmap_tailwind/config/routes.rb
@@ -1,5 +1,6 @@
 Rails.application.routes.draw do
   draw :pages
+  draw :errors
   draw :dev
   draw :support